Rename sem.rs -> semantics.rs
