+++ compiler-lib/src/lib.rs
@@ -25,6 +25,6 @@ pub mod parser;
 #[macro_use]
 pub mod visitor;
 pub mod print;
-pub mod sem;
+pub mod semantics;
 mod spantracker;
 pub mod strtab;
